Fixed bug in TextP.regex and TextP.notRegex serialization in Java
